refactor complex single-file app into multiple files in multiple packages
